+++ src/pages/google-oauth-callback/index.ts
@@ -1,2 +1,1 @@
 export { parseGoogleOAuthCallbackParams } from './model/parseGoogleOAuthCallbackParams'
-export { GoogleOAuthCallbackPage } from './ui/GoogleOAuthCallbackPage'